Move demo fetchers in a different file
